refactor: ASTToLLVMIRVisitor inherits from ASTToObjectVisitor
